rename(): LottoRepository main으로 변경
