+++ src/main/java/lotto/repository/LottoRepository.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.repository;
 
 import java.util.ArrayList;
 import java.util.List;
